Generalize matcher creation and restrict matcher string length
Since the matchers now all follow the same pattern we can get rid of
some boilerplate by creating a function which implements this pattern.

The builtin match function is terribly inefficient when given long
strings to work with; it even causes stack overflows when the string
is long enough. To alleviate this, introduce customizable max lengths
for all matchers.