Force evaluation of old state in readToken
For long enough files, we run into stack overflow issues, because Nix
holds on to old state. To fix this, force evaluation of old state for
every 1000 lines read; this seems like a good trade-off, where it
never has to be done for short enough files, but at the same time
doesn't degrade performance too much when evaluation has to be forced
and, going by my own emacs config, leaves plenty of headroom until the
state really _has_ to be forced to avoid disaster.