Simplify and improve flag parsing in parseOrgModeBabel using fromElisp
Instead of implementing a separate parsing for flags of Org mode babel
headers, parse them using `fromElisp`. The flags are more complex than
anticipated: the values can for example contain Elisp. Luckily, both
flags and values seem to be parsable as Elisp, so we can just pass the
rest of the line to `fromElisp` and get a nicely formatted list which
can easily be parsed as flags and values.